[#175589933] fix and refactoring
